Refactor APIs to use .json instead of .send to better work alongside code in frontend.
